% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_data.R
\name{cancer_genes}
\alias{cancer_genes}
\title{Available gene symbols in a cache (缓存中的全部基因)}
\usage{
cancer_genes(cache)
}
\arguments{
\item{cache}{cancer cache (癌种缓存)}
}
\value{
character vector (基因符号向量)
}
\description{
Available gene symbols in a cache (缓存中的全部基因)
}
