% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_combine.R
\name{default_basic_params}
\alias{default_basic_params}
\title{Default parameters for the basic dimension (default basic parameters)}
\usage{
default_basic_params()
}
\description{
Default parameters for the basic dimension (default basic parameters)
}
