% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_combine.R
\name{default_immune_params}
\alias{default_immune_params}
\title{Default parameters for the immune dimension (default immune parameters)}
\usage{
default_immune_params()
}
\description{
Default parameters for the immune dimension (default immune parameters)
}
