% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_combine.R
\name{default_metabolic_params}
\alias{default_metabolic_params}
\title{Default parameters for the metabolic dimension (default metabolic parameters)}
\usage{
default_metabolic_params()
}
\description{
Default parameters for the metabolic dimension (default metabolic parameters)
}
