% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_immune_meta.R
\name{.gene_trait_cor}
\alias{.gene_trait_cor}
\title{Gene-trait spearman correlation matrix (閸╁搫娲?閻楃懓绶?Spearman 閻╃鍙ч惌鈺呮█)
Uses pre-computed values for background genes and computes on the fly
for user-supplied genes. 閼冲本娅欓崺鍝勬礈閻劑顣╃拋锛勭暬閸婄》绱濋悽銊﹀煕閸╁搫娲滈崡铏鐠侊紕鐣婚妴?#' @noRd}
\usage{
.gene_trait_cor(cache, genes, trait_df, pre_mat)
}
\description{
Gene-trait spearman correlation matrix (閸╁搫娲?閻楃懓绶?Spearman 閻╃鍙ч惌鈺呮█)
Uses pre-computed values for background genes and computes on the fly
for user-supplied genes. 閼冲本娅欓崺鍝勬礈閻劑顣╃拋锛勭暬閸婄》绱濋悽銊﹀煕閸╁搫娲滈崡铏鐠侊紕鐣婚妴?#' @noRd
}
