% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_advanced.R
\name{gene_expr_log2}
\alias{gene_expr_log2}
\title{log2(x+1) gene expression for display (log2(x+1) 显示尺度)}
\usage{
gene_expr_log2(cache, gene, unit = c("tpm", "count", "fpkm"))
}
\description{
log2(x+1) gene expression for display (log2(x+1) 显示尺度)
}
