% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_data.R
\name{immune_traits}
\alias{immune_traits}
\title{Immune trait names for a method (某免疫浸润方法下的细胞特征名)}
\usage{
immune_traits(cache, method = "timer")
}
\description{
Immune trait names for a method (某免疫浸润方法下的细胞特征名)
}
