% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_immune_meta.R
\name{metabolic_cor}
\alias{metabolic_cor}
\title{Metabolic pathway correlation (代谢途径相关矩阵)}
\usage{
metabolic_cor(cache, genes)
}
\arguments{
\item{cache}{cancer cache (癌种缓存)}

\item{genes}{gene symbols (基因符号)}
}
\value{
matrix genes x pathways (基因 x 代谢途径的相关系数矩阵)
}
\description{
Metabolic pathway correlation (代谢途径相关矩阵)
}
