% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_data.R
\name{metabolic_traits}
\alias{metabolic_traits}
\title{Metabolic pathway names (代谢途径名称)}
\usage{
metabolic_traits(cache)
}
\arguments{
\item{cache}{cancer cache (癌种缓存)}
}
\value{
character vector of GSVA metabolic pathway names (代谢途径名称)
}
\description{
Metabolic pathway names (代谢途径名称)
}
