% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_expr_box}
\alias{plot_expr_box}
\title{Tumor vs normal expression boxplot}
\usage{
plot_expr_box(cache, gene)
}
\arguments{
\item{cache}{cancer cache}

\item{gene}{gene symbol}
}
\value{
a ggplot object
}
\description{
Tumor vs normal expression boxplot
}
