% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_advanced.R
\name{plot_facet_heatmap}
\alias{plot_facet_heatmap}
\title{Gene x cancer score heatmap, one facet per dimension (分面热图)}
\usage{
plot_facet_heatmap(
  d,
  score_col = "Score",
  dim_col = "Dim",
  gene_col = "Gene",
  cancer_col = "Cancer"
)
}
\arguments{
\item{d}{long data.frame with columns Gene, Cancer, Dim, Score}

\item{score_col}{column name of the score (评分列名)}

\item{dim_col}{column name of the dimension (维度列名)}

\item{gene_col}{column name of the gene (基因列名)}

\item{cancer_col}{column name of the cancer (癌种列名)}
}
\value{
a ggplot object
}
\description{
Gene x cancer score heatmap, one facet per dimension (分面热图)
}
