% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_immune_cor}
\alias{plot_immune_cor}
\title{Immune cell infiltration correlation plot}
\usage{
plot_immune_cor(cache, gene, method = "timer")
}
\arguments{
\item{cache}{cancer cache}

\item{gene}{gene symbol}

\item{method}{immune method: timer / cibersort / mcp / ssgsea / xcell}
}
\value{
a ggplot object
}
\description{
Immune cell infiltration correlation plot
}
