% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metabolic_cor}
\alias{plot_metabolic_cor}
\title{Metabolic pathway correlation plot}
\usage{
plot_metabolic_cor(cache, gene)
}
\arguments{
\item{cache}{cancer cache}

\item{gene}{gene symbol}
}
\value{
a ggplot object
}
\description{
Metabolic pathway correlation plot
}
