% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ranking}
\alias{plot_ranking}
\title{Ranking bar chart of a score column}
\usage{
plot_ranking(scores, value = "Combined_Score", top_n = 20)
}
\arguments{
\item{scores}{data.frame from score_genes}

\item{value}{column to rank by, e.g. "Combined_Score"}

\item{top_n}{number of genes to show}
}
\value{
a ggplot object
}
\description{
Ranking bar chart of a score column
}
