% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_roc_diag}
\alias{plot_roc_diag}
\title{Diagnostic ROC curve (tumor vs normal)}
\usage{
plot_roc_diag(cache, gene)
}
\arguments{
\item{cache}{cancer cache}

\item{gene}{gene symbol}
}
\value{
a ggplot object
}
\description{
Diagnostic ROC curve (tumor vs normal)
}
