% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_score_heatmap}
\alias{plot_score_heatmap}
\title{Gene x cancer score heatmap}
\usage{
plot_score_heatmap(scores_long, value = "Combined_Score")
}
\arguments{
\item{scores_long}{long table from score_genes_multi}

\item{value}{column to plot}
}
\value{
a ggplot object
}
\description{
Gene x cancer score heatmap
}
