% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer_data.R
\name{default_immune_weights}
\alias{default_immune_weights}
\title{Default immune weights per method (各免疫方法的默认权重)}
\usage{
default_immune_weights(method = "timer")
}
\description{
Default immune weights per method (各免疫方法的默认权重)
}
