% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_immune_meta.R
\name{metabolic_cor}
\alias{metabolic_cor}
\title{Metabolic pathway correlation (娴狅綀闃块柅鏂跨窞閻╃鍙ч幀褏鐓╅梼?}
\usage{
metabolic_cor(cache, genes)
}
\arguments{
\item{cache}{cancer cache (閻у瞼顫掔紓鎾崇摠)}

\item{genes}{gene symbols (閸╁搫娲滅粭锕€褰?}
}
\value{
matrix genes x pathways (閸╁搫娲?x 娴狅綀闃块柅鏂跨窞閻╃鍙ч惌鈺呮█)
}
\description{
Metabolic pathway correlation (娴狅綀闃块柅鏂跨窞閻╃鍙ч幀褏鐓╅梼?
}
